pyproject.toml
[tool.poetry]
name = "sonia"
version = "0.1.0"
description = ""
authors = ["An Long <aisk1988@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"backports.interpreters" = "^0.1.0"

[tool.poetry.dev-dependencies]
mypy = "^0.790"
flake8 = "^3.8.4"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
